move pe_tools to utility
