+++ unpacker/Unpacker.py
@@ -5,7 +5,7 @@ import time
 
 from unpacker.dumping.dumper import Dumper
 from unpacker.monitoring.monitor_manager import MonitorManager
-from unpacker.winwrapper.pe_tools import check_if_library
+from utility.pe_tools import check_if_library
 from unpacker.winwrapper.user_interaction import UserInteractor
 from unpacker.winwrapper.utilities import getUserPath, startAsLibrary, startAsExe
